refactor: split logoCircle into separate component
