+++ frontend/components/Homepage/CategoryCard/ProjectItem/index.vue
@@ -1,16 +1,7 @@
 <template>
   <div class="project">
     <div class="grid">
-      <div class="logo-circle" :style="{backgroundColor: color}">
-        <!--
-        TODO: Need to find an SVG letter font free for commercial use
-        ex: https://www.behance.net/gallery/48597451/Free-SVG-Alphabet
-                    -->
-        <!-- <img
-        class="logo-letter"
-        v-bind:src="`/src/assets/letters/${project.letter}.svg`"> -->
-        <div>{{ singleLetter }}</div>
-      </div>
+      <LogoCircle v-bind="$props" />
       <nuxt-link :to="link" class="project-title">
         {{ project.name }}
       </nuxt-link>
@@ -30,33 +21,15 @@
 import 'reflect-metadata'
 import { Vue, Component, Prop } from 'vue-property-decorator'
 import { Util } from '../../../../../shared/util'
-import { pickLogoLetter } from '../../../../assets/utils'
+import LogoCircle from './LogoCircle/index.vue'
 
-@Component
+@Component({
+  components: { LogoCircle }
+})
 export default class ProjectItemComponent extends Vue {
     @Prop() project! : any // TODO: Typings
     @Prop() projectIndex! : number
     @Prop() categoryIndex! : number
-
-    logoCircleColors = [
-      '#039BE5',
-      '#673AB7',
-      '#FBC02D',
-      '#FF7043',
-      '#C2185B',
-      '#009688',
-      '#9C27B0',
-      '#33AC71'
-    ]
-
-    get singleLetter () {
-      return pickLogoLetter(this.project.name)
-    }
-
-    get color () {
-      const colors = this.logoCircleColors
-      return colors[(this.projectIndex + this.categoryIndex) % colors.length]
-    }
 
     get description () {
       const words = this.project.description.split(' ')
@@ -119,33 +92,6 @@ export default class ProjectItemComponent extends Vue {
   float: left;
   position: relative;
   padding: 20px;
-}
-
-.logo-circle {
-  grid-area: logo;
-
-  align-self: center;
-
-  width: 50px;
-  height: 50px;
-  font-size: 25px;
-  line-height: 50px;
-  float: left;
-
-  margin-right: 4px;
-  margin-left: 0px;
-
-  border-radius: 100%;
-  display: grid;
-  justify-items: center;
-  align-items: center;
-  color: white;
-
-  img {
-    width: 74px;
-    height: 74px;
-    margin: 13px;
-  }
 }
 
 .project-title {
